+++ Assets/Scripts/MissileBehaviours.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Moving + Affecting Enemies + Player + UI:
+public class MissileBehaviour: MonoBehaviour {
+    public Vector2 movingSpeed;
+    public float movingTime;
+
+    GameObject explosion;
+    AudioSource flyingSound;
+    AudioSource explodingSound;
+
+
+}
+